# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "VSB 服务端 API 文档",
        "title": "VSB Server API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/permission/user/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建新用户，需要 Bearer Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "创建用户",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误或账号已存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/permission/user/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据用户 ID 删除用户，需要 Bearer Token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/permission/user/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "分页获取用户列表，支持 username/account 模糊搜索、email 精确匹配，需要 Bearer Token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "获取用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认 1",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认 20，上限 100",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户名（模糊搜索）",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账号（模糊搜索）",
                        "name": "account",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "邮箱（精确匹配）",
                        "name": "email",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.PageData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/permission/user/update/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据用户 ID 更新用户信息，空字段表示不修改，需要 Bearer Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "更新用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.StatusEnum": {
            "type": "string",
            "enum": [
                "active",
                "unlisted",
                "listed",
                "inactive",
                "deleted",
                "expired",
                "pending",
                "approved",
                "disabled",
                "enabled"
            ],
            "x-enum-comments": {
                "StatusActive": "活跃中",
                "StatusApproved": "审核通过",
                "StatusDeleted": "已删除",
                "StatusDisabled": "禁用",
                "StatusEnabled": "启用",
                "StatusExpired": "已过期",
                "StatusInactive": "已下架",
                "StatusListed": "已上架",
                "StatusPending": "待审核",
                "StatusUnlisted": "未上架"
            },
            "x-enum-descriptions": [
                "活跃中",
                "未上架",
                "已上架",
                "已下架",
                "已删除",
                "已过期",
                "待审核",
                "审核通过",
                "禁用",
                "启用"
            ],
            "x-enum-varnames": [
                "StatusActive",
                "StatusUnlisted",
                "StatusListed",
                "StatusInactive",
                "StatusDeleted",
                "StatusExpired",
                "StatusPending",
                "StatusApproved",
                "StatusDisabled",
                "StatusEnabled"
            ]
        },
        "response.PageData": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "当前页数据列表"
                },
                "pageIndex": {
                    "description": "当前页码，从 1 开始",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "total": {
                    "description": "总条数",
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据"
                },
                "message": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
        "user.UserRequest": {
            "type": "object",
            "required": [
                "account",
                "departmentId",
                "identity",
                "password",
                "roleId",
                "username"
            ],
            "properties": {
                "account": {
                    "description": "账号",
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间，由服务端生成",
                    "type": "string"
                },
                "departmentId": {
                    "description": "部门ID",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "identity": {
                    "description": "身份标识",
                    "type": "string",
                    "enum": [
                        "admin",
                        "user"
                    ]
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "roleId": {
                    "description": "角色ID",
                    "type": "string"
                },
                "status": {
                    "description": "用户状态",
                    "enum": [
                        "enabled",
                        "disabled"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.StatusEnum"
                        }
                    ]
                },
                "updatedAt": {
                    "description": "更新时间，由服务端生成",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "user.UserResponse": {
            "type": "object",
            "required": [
                "account",
                "createdAt",
                "departmentId",
                "identity",
                "roleId",
                "status",
                "updatedAt",
                "userId",
                "username"
            ],
            "properties": {
                "account": {
                    "description": "账号",
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "departmentId": {
                    "description": "部门ID",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "identity": {
                    "description": "身份标识",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                },
                "roleId": {
                    "description": "角色ID",
                    "type": "string"
                },
                "status": {
                    "description": "用户状态",
                    "enum": [
                        "enabled",
                        "disabled"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.StatusEnum"
                        }
                    ]
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "userId": {
                    "description": "用户ID",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "user.UserUpdateRequest": {
            "type": "object",
            "required": [
                "departmentId",
                "identity",
                "roleId",
                "username"
            ],
            "properties": {
                "departmentId": {
                    "description": "部门ID",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "identity": {
                    "description": "身份标识",
                    "type": "string"
                },
                "password": {
                    "description": "登录密码",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                },
                "roleId": {
                    "description": "角色ID",
                    "type": "string"
                },
                "status": {
                    "description": "用户状态",
                    "enum": [
                        "enabled",
                        "disabled"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.StatusEnum"
                        }
                    ]
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer JWT token，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}